+++ src/Parser/AbstractFileReader.php
@@ -7,7 +7,7 @@ namespace SugiPHP\Config\Parser;
 use SugiPHP\Config\Exception\FileException;
 
 /**
- * A parser should convert a string or a file into an associative array which
+ * A parser should convert data or a file into an associative array which
  * then can be passed to a Config class.
  */
 abstract class AbstractFileReader implements ParserInterface
@@ -15,12 +15,12 @@ abstract class AbstractFileReader implements ParserInterface
     /**
      * {@inheritDoc}
      */
-    abstract public function fromString(string $string): array;
+    abstract public function parse(mixed $data): array;
 
     /**
      * {@inheritDoc}
      */
-    public function fromFile(string $fileName): array
+    public function parseFile(string $fileName): array
     {
         // check if the $fileName is a real file and load it
         if (!is_file($fileName)) {
@@ -34,6 +34,42 @@ abstract class AbstractFileReader implements ParserInterface
             throw new FileException("Could not read configuration file {$fileName}");
         }
 
-        return $this->fromString($contents);
+        return $this->parse($contents);
+    }
+
+    /**
+     * @deprecated since 2.0.0, use parse() instead. Will be removed in a
+     *             future version.
+     */
+    public function fromString(string $string): array
+    {
+        @trigger_error(
+            sprintf(
+                '%s::fromString() is deprecated since 2.0.0, use %s::parse() instead. It will be removed in a future version.',
+                static::class,
+                static::class
+            ),
+            E_USER_DEPRECATED
+        );
+
+        return $this->parse($string);
+    }
+
+    /**
+     * @deprecated since 2.0.0, use parseFile() instead. Will be removed in a
+     *             future version.
+     */
+    public function fromFile(string $fileName): array
+    {
+        @trigger_error(
+            sprintf(
+                '%s::fromFile() is deprecated since 2.0.0, use %s::parseFile() instead. It will be removed in a future version.',
+                static::class,
+                static::class
+            ),
+            E_USER_DEPRECATED
+        );
+
+        return $this->parseFile($fileName);
     }
 }